+++ RELEASES.md
@@ -8,7 +8,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * service-template: hash of record, post-processing of record
 * service-template: browsing tab, with links to hash of record and template
 * prov-template-compiler: fixed issue with composite, post processing
-
+* service-template: hash for composite records
 
 `2.0.7`
 * prov-jsonld: fixed context url